+++ src/main/java/me/comu/exeter/events/CreateAChannelEvent.java
@@ -4,6 +4,7 @@ import me.comu.exeter.commands.admin.CreateAChannelCommand;
 import me.comu.exeter.logging.Logger;
 import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.events.guild.voice.GuildVoiceJoinEvent;
+import net.dv8tion.jda.api.events.guild.voice.GuildVoiceMoveEvent;
 import net.dv8tion.jda.api.events.guild.voice.GuildVoiceUpdateEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 
@@ -36,8 +37,7 @@ public class CreateAChannelEvent extends ListenerAdapter {
     @Override
     public void onGuildVoiceUpdate(@Nonnull GuildVoiceUpdateEvent event) {
         if (CreateAChannelCommand.isSet && !event.getEntity().getUser().isBot()) {
-            if (event.getChannelLeft() == null && event.getChannelJoined() != null)
-            {
+            if (event.getChannelLeft() == null && event.getChannelJoined() != null) {
                 if (event.getChannelJoined().getId().equalsIgnoreCase(CreateAChannelCommand.channelID)) {
                     Objects.requireNonNull(event.getJDA().getGuildById(CreateAChannelCommand.guildID)).createCopyOfChannel(Objects.requireNonNull(Objects.requireNonNull(event.getJDA().getGuildById(CreateAChannelCommand.guildID)).getVoiceChannelById(CreateAChannelCommand.channelID))).
                             setUserlimit(6).
@@ -52,18 +52,44 @@ public class CreateAChannelEvent extends ListenerAdapter {
                             }));
                 }
             }
-            if (event.getChannelJoined() == null && event.getChannelLeft() != null && CreateAChannelCommand.map.containsValue(event.getChannelLeft().getId())) {
-                if (event.getChannelLeft() != null && event.getChannelLeft().getMembers().isEmpty()) {
-                    try {
-                        CreateAChannelCommand.map.remove(event.getEntity().getId());
-                        event.getChannelLeft().delete().reason("Create-A-Channel VC").queue();
-                        Objects.requireNonNull(event.getJDA().getVoiceChannelById(CreateAChannelCommand.channelID)).putPermissionOverride(event.getEntity()).setAllow(Permission.VOICE_CONNECT).queue();
-                    } catch (Exception ex) {
-                        Logger.getLogger().print("Tried deleting an already deleted channel");
-                    }
+            if (event.getChannelLeft() != null && CreateAChannelCommand.map.containsValue(event.getChannelLeft().getId()) && event.getChannelLeft().getMembers().isEmpty()) {
+                try {
+                    CreateAChannelCommand.map.remove(event.getEntity().getId());
+                    event.getChannelLeft().delete().reason("Create-A-Channel VC").queue();
+                    Objects.requireNonNull(event.getJDA().getVoiceChannelById(CreateAChannelCommand.channelID)).putPermissionOverride(event.getEntity()).setAllow(Permission.VOICE_CONNECT).queue();
+                } catch (Exception ex) {
+                    Logger.getLogger().print("Tried deleting an already deleted channel");
                 }
+            }
+        }
+    }
 
+    @Override
+    public void onGuildVoiceMove(@Nonnull GuildVoiceMoveEvent event) {
+        if (CreateAChannelCommand.isSet && !event.getEntity().getUser().isBot()) {
+            if (event.getChannelJoined().getId().equalsIgnoreCase(CreateAChannelCommand.channelID)) {
+                Objects.requireNonNull(event.getJDA().getGuildById(CreateAChannelCommand.guildID)).createCopyOfChannel(Objects.requireNonNull(Objects.requireNonNull(event.getJDA().getGuildById(CreateAChannelCommand.guildID)).getVoiceChannelById(CreateAChannelCommand.channelID))).
+                        setUserlimit(6).
+                        setName(event.getEntity().getEffectiveName() + "'s Channel").
+                        reason("Create-A-Channel VC").
+                        queue((voiceChannel ->
+                        {
+                            Objects.requireNonNull(event.getJDA().getGuildById(CreateAChannelCommand.guildID)).moveVoiceMember(event.getEntity(), voiceChannel).queue();
+                            CreateAChannelCommand.map.put(event.getEntity().getId(), voiceChannel.getId());
+                            voiceChannel.upsertPermissionOverride(event.getEntity()).setAllow(Permission.MANAGE_CHANNEL).queue();
+                            Objects.requireNonNull(event.getJDA().getGuildById(CreateAChannelCommand.guildID).getVoiceChannelById(CreateAChannelCommand.channelID)).upsertPermissionOverride(event.getEntity()).setDeny(Permission.VOICE_CONNECT).queue();
+                        }));
+            }
+            if (event.getChannelLeft().getMembers().isEmpty() && CreateAChannelCommand.map.containsValue(event.getChannelLeft().getId())) {
+                try {
+                    CreateAChannelCommand.map.remove(event.getEntity().getId());
+                    event.getChannelLeft().delete().reason("Create-A-Channel VC").queue();
+                    Objects.requireNonNull(event.getJDA().getVoiceChannelById(CreateAChannelCommand.channelID)).putPermissionOverride(event.getEntity()).setAllow(Permission.VOICE_CONNECT).queue();
+                } catch (Exception ex) {
+                    Logger.getLogger().print("Tried deleting an already deleted channel");
+                }
             }
         }
     }
 }
+
